% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addHierarchyToStat}
\alias{addHierarchyToStat}
\title{Adds to the statistics matrix}
\usage{
addHierarchyToStat(stat, hierarchy, cluster_info, fn = "mean")
}
\arguments{
\item{stat}{a statitics matrix (rows=genes, columns=names of hierarchy\link{1})}

\item{hierarchy}{a hierarchy character vector}

\item{cluster_info}{A cluster_info matrix}
}
\value{
a statistics matrix with additional columns for other hierarchy levels
}
\description{
Adds to the statistics matrix
}
