% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{angle}
\alias{angle}
\title{angle}
\usage{
angle(x, y)
}
\arguments{
\item{x}{vector of length 2}

\item{y}{vector of length 2}
}
\value{
atan2 angle
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
