% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{avgangle}
\alias{avgangle}
\title{avangle}
\usage{
avgangle(x, y)
}
\arguments{
\item{x}{}

\item{y}{}
}
\value{
av angle
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
