% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargeTaxonomy.R
\name{chargeTaxonomyFromStats}
\alias{chargeTaxonomyFromStats}
\title{Convenience wrapper: read a backed h5ad plus Python-precomputed stats and write CHARGE.RData}
\usage{
chargeTaxonomyFromStats(
  AIT.file,
  stats.dir,
  charge.file.name = "CHARGE.RData",
  subsample = 1e+08,
  weight.by = "cell",
  seed = 42
)
}
\description{
Convenience wrapper: read a backed h5ad plus Python-precomputed stats and write CHARGE.RData
}
