% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargeTaxonomy.R
\name{chargeTaxonomyHybrid}
\alias{chargeTaxonomyHybrid}
\title{Save an RData object for use with CHARGE app}
\usage{
chargeTaxonomyHybrid(
  AIT.file,
  stats.dir = file.path(tempdir(), "charge_stats"),
  python.script = NULL,
  charge.file.name = "CHARGE.RData",
  subsample = 1e+08,
  weight.by = "cell",
  seed = 42,
  python.exe = "python3"
)
}
\arguments{
\item{AIT.file}{The location of a reference taxonomy anndata object in AIT format.  Note that for this function the h5ad file must be local.}

\item{stats.dir}{A directory where python should write precomputed statistics from chargeTaxonomyHybrid for reading back in. The default is typically fine as thse statistics are not needed if the function runs properly.}

\item{python.script}{The default (NULL) looks in the R library for the correct script. In almost all cases, this should not be changed.}

\item{charge.file.name}{File name (and path) to write CHARGE file to}

\item{subsample}{How much subsampling should be done before running statistics. Default (recommended!) is none, but subsampling can be done to speed up calculations.}

\item{weight.by}{Should statistics for higher levels of the hierarchy be weighted by "cell" (default), whereby each statistic is recalculated on all the cells in a given group for each level, or by "cluster", whereby statistics are averaged, counting each item from the first level of the hierarchy (e.g., cluster) evenly.}

\item{seed}{The seed to use for reproducibility}

\item{python.exe}{What is the name of the executable for python (Default is 'python3')}
}
\description{
This implementation of chargeTaxonomy uses python for all steps requiring reading data from the the X or raw.X slots (e.g., computing cluster statistics) and therefore should be compatible with large data sets.  \strong{Note that python needs to be properly install and referenced in your working environment in addition to the R docker environment below.}
}
\details{
Write an RData object with the following variables
\itemize{
\item counts_n - vector of total number of cells per cluster
\item counts - cell x cluster matrix where each value represents the number of cells in a cluster with at least one read for each gene
\item props - cell x cluster matrix where each value represents the fraction of cells in a cluster with at least one read for each gene
\item sums - cell x cluster matrix where each value represents the total number of counts of that gene for that cluster
\item means - cell x cluster matrix where each value represents the average counts of that gene for that cluster
\item sds - cell x cluster matrix where each value represents the standard deviation of counts of that gene for that cluster
\item hierarchy - cell type hierarchy stored in uns$hierarchy, but converted to a corrected ordered character vector
\item cluster_info - a table of cluster information that encodes the hierarchy for building sunburst plots
\item constellation - a list of plotly objects holding the constellation diagrams for each levels of the hierarchy
\strong{NOTE: This function is tested on "docker://alleninst/scrattch:1.1.4.1". We strongly encourage using this docker environment.}
}
}
\examples{
\dontrun{
  AIT.url  <- "https://allen-cell-type-taxonomies.s3.us-west-2.amazonaws.com/Human_MTG_SMART_seq_08082025.h5ad"
  AIT.file <- "Human_MTG_SMART_seq_08082025.h5ad"
  
  if (!file.exists(AIT.file)) {
    options(timeout = 9000)
    download.file(AIT.url, destfile = AIT.file, mode = "wb", timeout = 9000)
  }
  
  chargeTaxonomyHybrid(AIT.file)
}

}
