% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{edgeMaker}
\alias{edgeMaker}
\title{function to draw (curved) line between two points}
\usage{
edgeMaker(whichRow, len = 100, line.segments, curved = FALSE)
}
\value{
edge list
}
\description{
function to draw (curved) line between two points
}
