% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_knn_graph}
\alias{get_knn_graph}
\title{get knn graph}
\usage{
get_knn_graph(
  rd.dat,
  cl,
  k = 50,
  ref.cells = row.names(rd.dat),
  method = "Annoy.Cosine",
  knn.outlier.th = 2,
  outlier.frac.th = 0.5,
  clean.cells = row.names(rd.dat),
  knn.result = NULL,
  mc.cores = 10
)
}
\arguments{
\item{rd.dat}{Reduced dimension data frame where rows are cells and columns are reduced dimensional data (e.g., 50-100 principal components)}

\item{cl}{Vector of cluster assignments (can be numeric, character, or factor)}

\item{k}{Number of nearest neighbors (default=50 to align with example)}

\item{...}{Other variables that I'm not sure what they are and should probably be left alone}
}
\value{
A list with three components: knn.result = the knn result; knn.cl.df = the data frame of cluster information, outlier = outlier information.
}
\description{
This function gets a knn graph, which is required for creating a constellation diagram
}
