% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{perp}
\alias{perp}
\title{perp}
\usage{
perp(x, y, len, a, mid)
}
\arguments{
\item{x}{}

\item{y}{}

\item{len}{}

\item{a}{}

\item{mid}{}
}
\value{
perp value
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
