% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{perpEnd}
\alias{perpEnd}
\title{perpEnd}
\usage{
perpEnd(x, y, len)
}
\arguments{
\item{x}{}

\item{y}{}

\item{len}{}
}
\value{
perp at end
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
