% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{perpMid}
\alias{perpMid}
\title{perpMid}
\usage{
perpMid(x, y, len)
}
\arguments{
\item{x}{}

\item{y}{}

\item{len}{}
}
\value{
angle at midpoint
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
