% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellation.R
\name{perpStart}
\alias{perpStart}
\title{perpStart}
\usage{
perpStart(x, y, len)
}
\value{
perp start value
}
\description{
utils from vwline (https://github.com/pmur002/vwline) to draw variable width lines.
}
