% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refactorize_annotations}
\alias{refactorize_annotations}
\title{Reorder annotations to have the same order and colors as what is shown in a separate metadata file}
\usage{
refactorize_annotations(anno, metadata)
}
\arguments{
\item{anno}{an existing annotation data frame on which auto_annotate has ALREADY BEEN RUN}

\item{metadata}{a metadata file that includes some subset of content in the "_label" columns}
}
\value{
an updated data frame where order and colors of relevant columns have been adjusted to match metadata.
}
\description{
Reorder annotations to have the same order and colors as what is shown in a separate metadata file
}
