% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chargeTaxonomy.R
\name{chargeTaxonomy}
\alias{chargeTaxonomy}
\title{Save an RData object for use with CHARGE app}
\usage{
chargeTaxonomy(
  AIT.anndata,
  check.taxonomy = TRUE,
  subsample = 1e+08,
  charge.file.name = paste0(AIT.anndata$title, "_CHARGE.RData", seed = 42)
)
}
\arguments{
\item{AIT.anndata}{A reference taxonomy anndata object.}

\item{check.taxonomy}{Should the function check if the input variable is a valid AIT taxonomy (default = TRUE)}

\item{subsample}{How much subsampling should be done before running statistics. Default (recommended!) is none, but subsampling can be done to speed up calculations.}

\item{charge.file.name}{File name (and path) to write CHARGE file to}

\item{seed}{The seed to use for reproducibility}
}
\description{
Write an RData object with the following variables
\itemize{
\item counts_n - vector of total number of cells per cluster
\item counts - cell x cluster matrix where each value represents the number of cells in a cluster with at least one read for each gene
\item props - cell x cluster matrix where each value represents the fraction of cells in a cluster with at least one read for each gene
\item sums - cell x cluster matrix where each value represents the total number of counts of that gene for that cluster
\item means - cell x cluster matrix where each value represents the average counts of that gene for that cluster
\item sds - cell x cluster matrix where each value represents the standard deviation of counts of that gene for that cluster
\item hierarchy - cell type hierarchy stored in uns$hierarchy, but converted to a corrected ordered character vector
\item cluster_info
NOTE: proportions are not provided because they can be calculated trivaially from the above variables
}
}
\examples{
\dontrun{
  AIT.file    <- "HMBA_BG_Macaque_AIT.h5ad" #"Macaque_HMBA_basalganglia_AIT_pre-print.h5ad"
  AIT.anndata <- loadTaxonomy(AIT.file)
  #raw = AIT.anndata$raw$X
  #raw@x <- round(2^(raw@x)-1)
  #AIT.anndata$raw$X@x <- raw@x  # This does not work
  chargeTaxonomy(AIT.anndata)
}

}
